Move os components da página notes para uma pasta própria
