+++ app/app/notes/components/textarea.tsx
